replaced locateNames
findNames is the new replacement, and is far more suited for the task.

From a nested object this function will retrieve all key - value pairs, where the key matches with one of the given keywords.For my use, that will always be "navn".However, this is a general algorithm and can be applied otherwise.

    However, this general algorithm is wrapped inside a function that also takes all the names found, and all names that are not structured like { "fornavn": x, "etternavn": y }. Because of that, this module isn't general anymore.