minor bug fixes
sometimes when a pep was correctly labeled and only had one hit, the program would provide a summary of the same name that was searched.

this is because sometimes when you do a search for a part of a name, and only one result is found, then we have found one pep.